+++ packages/imandrax-api-models/src/imandrax_api_models/proto_models/artmsg.py
@@ -26,5 +26,5 @@ class Art(BaseModel):
         )
     )
     storage: list[StorageEntry] = Field(
-        default_factory=lambda: [], description='Additional definitions on the side'
+        default_factory=list, description='Additional definitions on the side'
     )